Fix semantic markup of fieldsets
Currently the page is constructed using nested `fieldset`s – although this may be syntactically correct it might produce difficulties semantically. Also, every `fieldset` requires a `legend`.

Replacing the outer `fieldset` with a `div` and the initial `legend` with an `h2` might make more sense semantically and reduces the complexity of the markup.